requirements.txt
aiofiles==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
chardet==5.1.0
geocoder==1.38.1
httpx==0.24.1
lxml==4.9.3
open-interpreter==0.1.4
openai==0.27.8
Pygments==2.15.1
pyperclip==1.8.2
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
qdrant-client==1.4.0
semantic-kernel==0.3.11.dev0
textual==0.37.1
tiktoken==0.4.0
unstructured==0.10.18
